requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
click==8.1.3
    # via flask
dataclasses==0.8
    # via werkzeug
flask==2.2.3
    # via -r requirements.in
importlib-metadata==6.1.0
    # via click
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via jinja2
psycopg2-binary==2.9.5
    # via -r requirements.in
pymysql==1.0.2
    # via -r requirements.in
redis==4.5.3
    # via -r requirements.in
typing-extensions==4.5.0
    # via importlib-metadata
werkzeug==2.2.3
    # via flask
zipp==3.15.0
    # via importlib-metadata
